+++ base/src/main/java/io/github/sds100/keymapper/base/constraints/ConstraintSnapshot.kt
@@ -5,6 +5,7 @@ import android.os.Build
 import io.github.sds100.keymapper.base.system.accessibility.IAccessibilityService
 import io.github.sds100.keymapper.common.utils.Orientation
 import io.github.sds100.keymapper.common.utils.PhysicalOrientation
+import io.github.sds100.keymapper.common.utils.SizeKM
 import io.github.sds100.keymapper.common.utils.firstBlocking
 import io.github.sds100.keymapper.system.bluetooth.BluetoothDeviceInfo
 import io.github.sds100.keymapper.system.camera.CameraAdapter
@@ -51,6 +52,7 @@ class LazyConstraintSnapshot(
         displayAdapter.cachedPhysicalOrientation
     }
     private val isScreenOn: Boolean by lazy { displayAdapter.isScreenOn.firstBlocking() }
+    private val displaySize: SizeKM by lazy { displayAdapter.size }
     private val appsPlayingMedia: List<String> by lazy {
         mediaAdapter.getActiveMediaSessionPackages()
     }
@@ -134,6 +136,19 @@ class LazyConstraintSnapshot(
 
             is ConstraintData.ScreenOff -> !isScreenOn
             is ConstraintData.ScreenOn -> isScreenOn
+
+            // Compare the resolution regardless of orientation so that the constraint
+            // holds whether the device is in portrait or landscape.
+            is ConstraintData.DisplayResolution ->
+                (
+                    displaySize.width == constraint.data.width &&
+                        displaySize.height == constraint.data.height
+                    ) ||
+                    (
+                        displaySize.width == constraint.data.height &&
+                            displaySize.height == constraint.data.width
+                        )
+
             is ConstraintData.FlashlightOff -> !cameraAdapter.isFlashlightOn(constraint.data.lens)
             is ConstraintData.FlashlightOn -> cameraAdapter.isFlashlightOn(constraint.data.lens)
             is ConstraintData.WifiConnected -> {