+++ base/src/main/java/io/github/sds100/keymapper/base/sorting/comparators/KeyMapConstraintsComparator.kt
@@ -115,6 +115,9 @@ class KeyMapConstraintsComparator(
             is ConstraintData.RingerMode -> Success("")
             is ConstraintData.ScreenOff -> Success("")
             is ConstraintData.ScreenOn -> Success("")
+            is ConstraintData.DisplayResolution -> Success(
+                "${constraint.data.width}x${constraint.data.height}",
+            )
             is ConstraintData.WifiConnected -> if (constraint.data.ssid == null) {
                 Success("")
             } else {