+++ base/src/test/java/io/github/sds100/keymapper/base/utils/TestConstraintSnapshot.kt
@@ -5,6 +5,7 @@ import io.github.sds100.keymapper.base.constraints.ConstraintData
 import io.github.sds100.keymapper.base.constraints.ConstraintSnapshot
 import io.github.sds100.keymapper.common.utils.Orientation
 import io.github.sds100.keymapper.common.utils.PhysicalOrientation
+import io.github.sds100.keymapper.common.utils.SizeKM
 import io.github.sds100.keymapper.system.bluetooth.BluetoothDeviceInfo
 import io.github.sds100.keymapper.system.camera.CameraLens
 import io.github.sds100.keymapper.system.foldable.HingeState
@@ -36,6 +37,7 @@ class TestConstraintSnapshot(
     val localTime: LocalTime = LocalTime.now(),
     val hingeState: HingeState = HingeState.Unavailable,
     val isNotificationPanelShowing: Boolean = false,
+    val displaySize: SizeKM = SizeKM(1080, 1920),
 ) : ConstraintSnapshot {
 
     override fun isSatisfied(constraint: Constraint): Boolean {
@@ -72,6 +74,9 @@ class TestConstraintSnapshot(
 
             is ConstraintData.ScreenOff -> !isScreenOn
             is ConstraintData.ScreenOn -> isScreenOn
+            is ConstraintData.DisplayResolution ->
+                (displaySize.width == data.width && displaySize.height == data.height) ||
+                    (displaySize.width == data.height && displaySize.height == data.width)
             is ConstraintData.FlashlightOff -> when (data.lens) {
                 CameraLens.BACK -> !isBackFlashlightOn
                 CameraLens.FRONT -> !isFrontFlashlightOn